When in shutdown state all mailbox transactions are dropped and fatal/non-fatal error registers inform the SoC that Caliptra services have been shutdown.
